separate deposit contract compiler and tester, pin compiler to python 3.7. Upgrade others to 3.8
