--- deposit_contract/requirements-testing.txt
@@ -1,5 +1,0 @@
-eth-tester[py-evm]==0.1.0b39
-git+https://github.com/vyperlang/vyper@1761-HOTFIX-v0.1.0-beta.13
-web3==5.0.0b2
-pytest==3.6.1
-../tests/core/pyspec
