--- deposit_contract/tests/contracts/test_compile.py
@@ -1,19 +1,0 @@
-from vyper import (
-    compiler,
-)
-
-from .utils import (
-    get_deposit_contract_code,
-    get_deposit_contract_json,
-)
-
-
-def test_compile_deposit_contract():
-    compiled_deposit_contract_json = get_deposit_contract_json()
-
-    deposit_contract_code = get_deposit_contract_code()
-    abi = compiler.mk_full_signature(deposit_contract_code)
-    bytecode = compiler.compile_code(deposit_contract_code)['bytecode']
-
-    assert abi == compiled_deposit_contract_json["abi"]
-    assert bytecode == compiled_deposit_contract_json["bytecode"]
